refactor: Move log_dirty_pages to the module where it's used
